Add files via upload
Chess is the main file to run.  Game_selector is called to allow you to load old files, and also has function to display which pieces have been taken from the board.
to show chess pieces you need the ARIALUNI font.  I tried a few other fonts and it would not do the chess pieces on my MacOS.