Add files via upload
Updated to USE a custom LLM as one of the chess engines.  Also ability to play different models against each other, or against search type methods.  Used the 64x63 potential moves on a chess board as the tokenizer. Trained the chess transformer to on Stockfish games.  Usually 2 million or so games for 1 or 1.5 epochs